client: fix signed/unsigned compilation warnings

+++ src/client/Client.cc
@@ -6279,10 +6279,10 @@ Fh *Client::_create_fh(Inode *in, int flags, int cmode)
   f->readahead.set_min_readahead_size(conf->client_readahead_min);
   uint64_t max_readahead = Readahead::NO_LIMIT;
   if (conf->client_readahead_max_bytes) {
-    max_readahead = MIN(max_readahead, conf->client_readahead_max_bytes);
+    max_readahead = MIN(max_readahead, (uint64_t)conf->client_readahead_max_bytes);
   }
   if (conf->client_readahead_max_periods) {
-    max_readahead = MIN(max_readahead, conf->client_readahead_max_periods * p);
+    max_readahead = MIN(max_readahead, ((uint64_t)conf->client_readahead_max_periods) * p);
   }
   f->readahead.set_max_readahead_size(max_readahead);
   vector<uint64_t> alignments;
